Add Q1 and Q3 aggregates
