+++ cpp/perspective/src/include/perspective/base.h
@@ -259,6 +259,8 @@ enum t_aggtype {
     AGGTYPE_WEIGHTED_MEAN,
     AGGTYPE_UNIQUE,
     AGGTYPE_ANY,
+    AGGTYPE_Q1,
+    AGGTYPE_Q3,
     AGGTYPE_MEDIAN,
     AGGTYPE_JOIN,
     AGGTYPE_SCALED_DIV,
