+++ cpp/perspective/src/include/perspective/sparse_tree.h
@@ -312,7 +312,8 @@ public:
 
     t_tscalar get_aggregate(t_index idx, t_index aggnum) const;
 
-    t_tscalar get_aggregate_median(std::vector<t_tscalar>& values) const;
+    t_tscalar
+    get_aggregate_median(std::vector<t_tscalar>& values, float n) const;
 
     void get_child_indices(t_index idx, std::vector<t_index>& out_data) const;
